+++ src/store/socket.js
@@ -172,6 +172,11 @@ class LiveSession {
         if (!this._isSpectator) return;
         this._store.commit("toggleNight", params);
         break;
+      case "isVoteHistoryAllowed":
+        if (!this._isSpectator) return;
+        this._store.commit("session/setVoteHistoryAllowed", params);
+        this._store.commit("session/clearVoteHistory");
+        break;
       case "votingSpeed":
         if (!this._isSpectator) return;
         this._store.commit("session/setVotingSpeed", params);
@@ -268,6 +273,7 @@ class LiveSession {
       this._sendDirect(playerId, "gs", {
         gamestate: this._gamestate,
         isNight: grimoire.isNight,
+        isVoteHistoryAllowed: session.isVoteHistoryAllowed,
         nomination: session.nomination,
         votingSpeed: session.votingSpeed,
         lockedVote: session.lockedVote,
@@ -289,6 +295,7 @@ class LiveSession {
       gamestate,
       isLightweight,
       isNight,
+      isVoteHistoryAllowed,
       nomination,
       votingSpeed,
       votes,
@@ -340,6 +347,7 @@ class LiveSession {
     });
     if (!isLightweight) {
       this._store.commit("toggleNight", !!isNight);
+      this._store.commit("session/setVoteHistoryAllowed", isVoteHistoryAllowed);
       this._store.commit("session/nomination", {
         nomination,
         votes,
@@ -687,6 +695,17 @@ class LiveSession {
   }
 
   /**
+   * Send the isVoteHistoryAllowed state. ST only
+   */
+  setVoteHistoryAllowed() {
+    if (this._isSpectator) return;
+    this._send(
+      "isVoteHistoryAllowed",
+      this._store.state.session.isVoteHistoryAllowed
+    );
+  }
+
+  /**
    * Send the voting speed. ST only
    * @param votingSpeed voting speed in seconds, minimum 1
    */
@@ -839,6 +858,9 @@ export default store => {
         break;
       case "session/clearVoteHistory":
         session.clearVoteHistory();
+        break;
+      case "session/setVoteHistoryAllowed":
+        session.setVoteHistoryAllowed();
         break;
       case "toggleNight":
         session.setIsNight();